Autolinking now working, one package. But Appregistry error
