+++ example/index.js
@@ -1,5 +1,9 @@
+/**
+ * @format
+ */
+
 import { AppRegistry } from 'react-native';
 import App from './App';
 import { name as appName } from "./app.json";  
-//require('RCTNativeAppEventEmitter');
+
 AppRegistry.registerComponent(appName, () => App);
