--- lib/module/analytics/states.js
@@ -1,20 +1,0 @@
-// deviceinfo.js
-'use strict';
-import { NativeModules, Platform } from 'react-native';
-
-const AccStates = NativeModules.RNAccStates;
-
-class States {
-
-    setState(name, value) {
-    	if (Platform.OS === 'ios') {
-        	AccStates.setState(value, name);
-    	} else {
-    		AccStates.setState(name, value);
-    	}
-    }
-}
-
-var states = new States();
-
-exports.states = states;
