--- lib/module/control.js
@@ -1,95 +1,0 @@
-// control.js
-'use strict';
-
-import { NativeModules, Platform } from 'react-native';
-
-const AccControl = NativeModules.RNAccControl;
-
-class Control {
-
-  setOptinDataEnabled(isEnabled) {
-    AccControl.setOptinDataEnabled(isEnabled);
-  }
-
-  setOptinGeolocEnabled(isEnabled) {
-    AccControl.setOptinGeolocEnabled(isEnabled);
-  }
-  
-  isOptinDataEnabled() {
-    return AccControl.isOptinDataEnabled();
-  }
-
-  isOptinGeolocEnabled() {
-    return AccControl.isOptinGeolocEnabled();
-  }
- 
-  setAllServicesEnabled(isEnabled) {
-    if (Platform.OS === 'ios') {
-        AccControl.setAllServicesEnabled(isEnabled);
-    }
-  }
-
-  areAllServicesEnabled() {
-    if (Platform.OS === 'ios') {
-      return AccControl.areAllServicesEnabled();
-    } else {
-      return new Promise(function (resolve) {
-        resolve(true);
-      });
-    }
-  }
-
-  setNetworkCallsEnabled(isEnabled) {
-    if (Platform.OS === 'ios') {
-      AccControl.setNetworkCallsEnabled(isEnabled);
-    }
-  }
-
-  areNetworkCallsEnabled() {
-    if (Platform.OS === 'ios') {
-      return AccControl.areNetworkCallsEnabled();
-    } else {
-      return new Promise(function (resolve) {
-        resolve(true);
-      });
-    }
-  }
-
-  setGeofenceServiceEnabled(isEnabled) {
-    if (Platform.OS === 'ios') {
-      AccControl.setGeofenceServiceEnabled(isEnabled);
-    }
-  }
-
-  isGeofenceServiceEnabled() {
-    if (Platform.OS === 'ios') {
-      return AccControl.isGeofenceServiceEnabled();
-    }
-    if (Platform.OS === 'android') {
-      return new Promise(function (resolve) {
-        resolve(true);
-      });
-    }
-  }
-
-  setBeaconServiceEnabled(isEnabled) {
-    if (Platform.OS === 'ios') {
-      AccControl.setBeaconServiceEnabled(isEnabled);
-    }
-  }
-
-  isBeaconServiceEnabled() {
-    if (Platform.OS === 'ios') {
-      return AccControl.isBeaconServiceEnabled();
-    }
-    if (Platform.OS === 'android') {
-      return new Promise(function (resolve) {
-        resolve(true);
-      });
-    }
-  }
-}
-
-var control = new Control();
-
-exports.control = control;
