--- lib/module/inbox.js
@@ -1,17 +1,0 @@
-// inbox.js
-'use strict';
-
-class Inbox {
-
-    getInbox() {
-
-    }
-
-    updateMessages(inbox) {
-
-    }
-}
-
-var inbox = new Inbox();
-
-exports.inbox = inbox;
