+++ rn-cli.config.js
@@ -1,0 +1,9 @@
+const blacklist = require('metro-config/src/defaults/blacklist');
+
+module.exports = {
+  resolver:{
+    blacklistRE: blacklist([
+            /node_modules\/.*\/node_modules\/react-native\/.*/,
+        ])
+  },
+};
